Reduce size of temp field in logs

+++ src/baro.hpp
@@ -5,6 +5,6 @@ void baro_setup();
 
 // Can return NaN if data hasn't been recieved from the barometer yet.
 float baro_get_altitude();
-int32_t baro_get_temp();
+int16_t baro_get_temp();
 
 void baro_print();
